backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
ytmusicapi==1.9.1
asyncpg==0.30.0
requests==2.32.3
python-dotenv==1.0.1
